+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Pending
 
+* Fix username change form to actually change the username (#327)
+
 ## Releases
 
 ### Deploy for 2018-03-15